Support for open and closed sensors
In sensors such as that of a door cover, on and off states are not as intuitive as open and closed. Additionally, in my case my custom sensors report status as open or closed so I made this change to permit specifying dynamic icons as can currently be done in on/off sensors.